// frontend/src/api/model/chunkRouting.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ChunkRouting {
    /** @pattern ^-?(?:0|[1-9]\d*)$ */
    chunkIndex?: number | string;
    category?: string;
    disposition?: string;
    /** @nullable */
    reason?: string | null;
    /** @nullable */
    sceneId?: string | null;
}