// frontend/src/api/model/chunkSummary.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */

export interface ChunkSummary {
    /** @pattern ^-?(?:0|[1-9]\d*)$ */
    index?: number | string;
    role?: string;
    category?: string;
    /** @pattern ^-?(?:0|[1-9]\d*)$ */
    chars?: number | string;
    head?: string;
}